requirements-dev.txt
-r requirements.txt

pytest==8.3.5
pytest-django==4.11.1